// apps/web/app/api/generated/schemas/getProjects200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Catalyst API
 * Type-safe REST API for the Catalyst project management system
 * OpenAPI spec version: 1.0.0
 */
import type { GetProjects200DataItemDescription } from './getProjects200DataItemDescription';
import type { GetProjects200DataItemStatus } from './getProjects200DataItemStatus';

export type GetProjects200DataItem = {
  id: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  description?: GetProjects200DataItemDescription;
  status: GetProjects200DataItemStatus;
  createdAt: string;
  updatedAt: string;
};